% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnfst_936K}
\alias{mnfst_936K}
\title{EPIC_v2 manifest file}
\format{
A data frame with 937691 rows and 5 columns:
\describe{
\item{Name}{CpG name}
\item{AddressA_ID}{AdressA ID}
\item{AddressB_ID}{AdressB ID}
\item{Infinium_Design_Type}{Infinium design type}
\item{Color_Channel}{Color channel}
}
}
\source{
\url{https://webdata.illumina.com}
}
\usage{
data(mnfst_936K)
}
\description{
A dataset containing all EPIC_v2 probes information.
}
\keyword{datasets}
