% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probelist_936K}
\alias{probelist_936K}
\title{Probe list for EPIC_v2}
\format{
A data frame with 62 rows and 3 columns:
\describe{
\item{CpG}{CpG list}
\item{Type}{Probe types}
\item{A2}{Alternative alleles}
}
}
\usage{
data(probelist_936K)
}
\description{
A dataset containing the list of 65 SNP probes.
}
\keyword{datasets}
